キャラクター作成画面の遷移とファイルの階層を整えた
