+++ lib/screens/character_creation_screen.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class CharacterCreationScreen extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('キャラクター作成'),
+      ),
+      body: Center(
+        child: Text('キャラクター作成画面'),
+      ),
+    );
+  }
+}
